First step to project manual syncing
So, I think I have it such that there is a flag that if true, will only
save things to the local DB, and if false will save to both the server
and DB. It at some point in the future it finds that an object is in the
DB, but not commited to the server, it will notice. Still don't have it
actually commiting the data, and need a button to trigger the checks as
well, but making headway.